ysfvender的头文件需要更新  之前遗漏了

+++ SDK/YSFVendor/Foundation/YSFUtil/YSFTimer.h
@@ -19,7 +19,7 @@ typedef void(^YSFTimerBlock)();
  *  @param repeats  是否重复
  *  @param cb  定时任务
  */
-- (void)start:(dispatch_queue_t)queue interval:(NSInteger)interval repeats:(BOOL)repeats block:(YSFTimerBlock)block;
+- (void)start:(dispatch_queue_t)queue interval:(double)interval repeats:(BOOL)repeats block:(YSFTimerBlock)block;
 
 /**
  *  开始定时
@@ -30,7 +30,7 @@ typedef void(^YSFTimerBlock)();
  *  @param repeats  是否重复
  *  @param cb  定时任务
  */
-- (void)start:(dispatch_queue_t)queue startAfter:(NSInteger)startAfter interval:(NSInteger)interval
+- (void)start:(dispatch_queue_t)queue startAfter:(double)startAfter interval:(NSInteger)interval
             repeats:(BOOL)repeats block:(YSFTimerBlock)block;
 
 /**
